All updates to aocl-3.2

+++ src/ref/atan2.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2021 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2008-2022 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
  * are permitted provided that the following conditions are met:
@@ -541,7 +541,7 @@ double ALM_PROTO_REF(atan2)(double y, double x)
 
   /* Find properties of arguments x and y. */
 
-  unsigned long long ux, ui, aux, xneg, uy, auy, yneg;
+  uint64_t ux, ui, aux, xneg, uy, auy, yneg;
 
   GET_BITS_DP64(x, ux);
   GET_BITS_DP64(y, uy);
