+++ src/ref/atan2f.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2021 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2008-2022 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
  * are permitted provided that the following conditions are met:
@@ -288,7 +288,7 @@ float ALM_PROTO_REF(atan2f)(float fy, float fx)
   double u, v, vbyu, q, s, uu, r;
   unsigned int swap_vu, index, xzero, yzero, xnan, ynan, xinf, yinf;
   int xexp, yexp, diffexp;
-  unsigned long long uy;
+  uint64_t uy;
 
   double x;
   double y;
